==> 00b_data_parse.ipynb <==
### Now that the data retreival is cleared up and we don't have any failed values or discrepencies, this notebook will allow the parsing from .xslx to .csv files that can be used in our modeling notebooks!

import matplotlib.pyplot as plt
import pandas as pd
import numpy as np

### To deal with .xlsx extension, you will need to install the necessary package with the following command:

*pip install openpyxl*

file_name = 'data/fixed_10gate.xlsx'
sheet_name = '10_Gates_1_run_FAILS_fixed'

def read_sheet(file_name, sheet_name): return pd.read_excel(file_name, sheet_name, engine='openpyxl')

data = read_sheet(file_name, sheet_name)

data.head(5)

param_strs = data[data.base.map(lambda x: len(x) > 10)].base.values
param_strs[:2]

trise_vals = data[data.base == 'T_RISE ']['values'].values
tfall_vals = data[data.base == 'T_FALL ']['values'].values
tdelay_vals = data[data.base == 'T_DELAY ']['values'].values
len(trise_vals) == len(tfall_vals) == len(tdelay_vals) == len(param_strs)

#trise_vals

#param_strs[0].split(sep='/')

It goes - [1] gate name, [2] clock cylce, [3] thickness, [4] voltage --> after some parsing

gate_types = pd.Series(param_strs).map(lambda x: x.split('/')).map(lambda x: x[1][2:])

clock_cycle = pd.Series(map(lambda x: x.split(sep='/')[2][:-2], param_strs))
thickness = pd.Series(map(lambda x: x.split(sep='/')[3][:-2], param_strs))
clock_cycle[0], thickness[0]

#how to remove the pesky 'p' value in the thickness
#thickness[0].replace('p', '.')

thickness = thickness.map(lambda x: x.replace('p', '.'))
thickness[:2]

voltage = pd.Series(map(lambda x: x.split('/')[-1].split('_')[-1], param_strs))

sample_clean_df = pd.DataFrame({'gate_types' : gate_types,
                        'voltage':voltage, 'thickness':thickness, 'clock_cycle': clock_cycle,
                        't_rise': trise_vals, 't_fall': tfall_vals, 't_delay': tdelay_vals})
sample_clean_df.tail(2)

def get_clean_df(df):
    param_strs = df[df.base.map(lambda x: len(x) > 10)].base.values
    gate_types = pd.Series(param_strs).map(lambda x: x.split('/')).map(lambda x: x[1][2:])
    trise_vals = df[df.base == 'T_RISE ']['values'].values
    tfall_vals = df[df.base == 'T_FALL ']['values'].values
    tdelay_vals = df[df.base == 'T_DELAY ']['values'].values
    clock_cycle = pd.Series(map(lambda x: x.split(sep='/')[2][:-2], param_strs))
    thickness = pd.Series(map(lambda x: x.split(sep='/')[3][:-2], param_strs))
    thickness = thickness.map(lambda x: x.replace('p', '.'))
    voltage = pd.Series(map(lambda x: x.split('/')[-1].split('_')[-1], param_strs))
    #pdb.set_trace()
    return pd.DataFrame({'type' : gate_types,
            'voltage':voltage, 'thickness':thickness, 'clock_cycle': clock_cycle,
            't_rise': trise_vals, 't_fall': tfall_vals, 't_delay': tdelay_vals})

run_df = get_clean_df(data).rename(columns={"type":"gate_type"})
run_df.shape

run_df.to_csv('data/ten_gate.csv', index=False)

run_df.head(1)

run_df.tail().t_delay.values

### FAILED

Lets make sense of the number of FAILED values in the runs

def get_failed(df, row_type='t_delay'):
    failed = (df[row_type] == ' FAILED').sum()
    return failed, f'{row_type}: {failed/len(df) * 100: .2f}%'
def get_full_fail(df):
    time_vals = ['t_delay', 't_rise', 't_fall']
    return [get_failed(df, t_val) for t_val in time_vals]

#get_failed(run1_df, 't_fall'), get_failed(run1_df, 't_rise'), get_failed(run1_df, 't_delay')

run_df[run_df.t_fall == ' FAILED'].head(2)

get_full_fail(run_df)

get_full_fail(run_df)

get_full_fail(run_df)

get_full_fail(run_df)

get_full_fail(run_df)

run_len = run_df.shape[0]
run_len

Just to check that we are getting 11k samples as expected!
