==> 01_random_forest.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('data/run1.csv')
df.head(2)

def get_gate(gate_type='buffer'): return df[df.gate_type == gate_type]

get_gate()

unique_gates = list(df.gate_type.unique())
unique_gates

def get_gate_xy(df, gate_type='buffer'):
    x = np.vstack([df.clock_cycle, df.thickness, df.voltage]).T
    y = np.vstack([df.t_delay, df.t_fall, df.t_rise]).T
    return x, y

x, y = get_gate_xy(df)
x.shape, y.shape

df[df.t_rise != ' FAILED'].shape, df[df.t_delay != ' FAILED'].shape, df.shape

gate_dict = dict.fromkeys(unique_gates)
gate_dict

for gate in unique_gates: 
    gate_dict[gate] = get_xy_df(df, gate_type=gate)

gate_dict['buffer'][0].shape

from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import Ridge
from sklearn.ensemble import RandomForestRegressor 
from sklearn.metrics import mean_squared_error, r2_score
from sklearn.model_selection import train_test_split

def ridge_reg(x_train, y_train, x_valid, y_valid):
    print('Doing Ridge Regression.....')
    ridge = Ridge(alpha = 1)
    ridge.fit(x_train, y_train)
    print('training score = ', ridge.score(x_train, y_train))
    y_pred_ridge = ridge.predict(x_valid)
    mse_ridge = mean_squared_error(y_valid, y_pred_ridge) 
    print('test score =', ridge.score(x_valid, y_valid)) 

def do_rf(x_train, y_train, x_valid, y_valid):
    print('\n\ndoing RF.....')
    RF = RandomForestRegressor(max_depth=100, random_state=0)
    RF.fit(x_train, y_train)
    print('trainig score = ', RF.score(x_train, y_train))
    y_pred = RF.predict(x_valid)
    mse = mean_squared_error(y_valid, y_pred)
    r2 = r2_score(y_valid, y_pred) # coefficient of determination
    print('test score =', RF.score(x_valid, y_valid))
    print(f'the mse is {mse}')
    print(f"the r2 score is: {r2}")
    return y_pred

def train_models(x, y, random_state=0):
    x_train, x_valid, y_train, y_valid = train_test_split(x, y, 
                                                        random_state=random_state, 
                                                        test_size=0.1)    
    ridge_reg(x_train, y_train, x_valid, y_valid)
    y_pred = do_rf(x_train, y_train, x_valid, y_valid)
    return y_pred

### Lets start with our Buffer circuit

_ = train_models(X_buffer, y_buffer, random_state=123)

### Inverter

_ = train_models(X_inverter, y_inverter, random_state=0)

### Nand

_ = train_models(X_nand, y_nand, random_state=1)

### Nor

_ = train_models(X_nor, y_nor)

# Normalizing the dependent variable helps?

Looks like things are not working as expected out of the box! Normally you do not need to do much with the data with a random-forest-classifier. but apparently this is not the case with regressors.....

_ = train_models(X_nor, y_nor/y_nor.min())

So by normalizing the dependent values we can replicate the previous Random Forest Model perforamnce....

_ = train_models(X_buffer, y_buffer/y_buffer.min())

_ = train_models(X_inverter, y_inverter / y_inverter.min())

_ = train_models(X_nand, y_nand / y_nand.min())
