==> 02_xgboost.ipynb <==
# XGBoost! 

## ...can we improve our random forest results?

"xgboost is all you need" --> https://arxiv.org/pdf/2110.01889.pdf (for tabular data)

from sklearn.metrics import mean_squared_error, r2_score
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor 
import xgboost as xgb
import pandas as pd
import numpy as np

df = pd.read_csv('final.csv')
types = ['inverter', 'buffer', 'nand', 'nor']
type_dict = dict(zip(df.type.unique(), types))
df.type = df.type.map(type_dict)
df.clock_cycle = df.clock_cycle.map(lambda x: float(x[:-2]))
df['thickness tFE'] = df['thickness tFE'].map(lambda x: float(x[:-2]))
drop_50ps_idx = np.where(df['clock_cycle'] == 50)
df.drop(index=drop_50ps_idx[0], inplace=True)
buffer_df = df[df.type == 'buffer']
inverter_df = df[df.type == 'inverter']
nor_df = df[df.type == 'nor']
nand_df = df[df.type == 'nand']


#df.iloc[:, 4:].sum(axis=1)
df.head(2)

#df.loc[[df.columns[-1]]]
#df.loc[df.columns[-1]]
df.loc[:, df.columns[4:]]

def get_xy_df(df):
    X_buffer = np.vstack([df.clock_cycle, df['thickness tFE'], df['Vdd']]).T
    y_buffer = np.vstack([df['T_RISE '], df['T_FALL '], df['T_DELAY ']]).T
    return X_buffer, y_buffer

X_buffer, y_buffer = get_xy_df(buffer_df)
X_inverter, y_inverter = get_xy_df(inverter_df)
X_nor, y_nor = get_xy_df(nor_df)
X_nand, y_nand = get_xy_df(nand_df)
X_buffer.shape, y_buffer.shape

regressor=xgb.XGBRegressor(eval_metric='rmse')

x_train, x_valid, y_train, y_valid = train_test_split(X_buffer, y_buffer, test_size=0.1)    

from sklearn.model_selection import GridSearchCV
# set up our search grid
param_grid = {"max_depth":    [4, 5],
              "n_estimators": [500, 600, 700],
              "learning_rate": [0.01, 0.015]}

# try out every combination of the above values
search = GridSearchCV(regressor, param_grid, cv=5).fit(x_train, y_train)

print("The best hyperparameters are ", search.best_params_)

#regressor.fit(X_buffer, y_buffer)

regressor=xgb.XGBRegressor(learning_rate = search.best_params_["learning_rate"],
                           n_estimators  = search.best_params_["n_estimators"],
                           max_depth     = search.best_params_["max_depth"],)

preds = regressor.predict(X_buffer)
preds.shape

regressor.score(X_buffer, y_buffer) ** 2, mean_squared_error(y_buffer,preds)

These two values above should be the same?

Lets keep our random forest for comparison

def do_rf(x_train, y_train, x_valid, y_valid):
    print('\n\ndoing RF.....')
    RF = RandomForestRegressor(max_depth=100, random_state=0)
    RF.fit(x_train, y_train)
    train_score = RF.score(x_train, y_train)
    y_pred = RF.predict(x_valid)
    mse = mean_squared_error(y_valid, y_pred)
    r2 = r2_score(y_valid, y_pred) # coefficient of determination
    valid_score = RF.score(x_valid, y_valid)

    print(f'the training score is {train_score:3f}')
    print(f'the validation score is {valid_score:3f}')
    print(f'the mse is {mse:3f}')
    print(f"the r2 score is: {r2:3f}")
    return y_pred

def do_xgb(x_train, y_train, x_valid, y_valid):
    print('\n\nboosting the trees.....')
    regressor=xgb.XGBRegressor(eval_metric='rmse')
    regressor.fit(x_train, y_train)
    train_score = regressor.score(x_train, y_train) ** 2 #bc using RMSE
    y_pred = regressor.predict(x_valid)
    mse = mean_squared_error(y_valid, y_pred)
    r2 = r2_score(y_valid, y_pred) # coefficient of determination
    valid_score = regressor.score(x_valid, y_valid)

    print(f'the training score is {train_score:3f}')    
    print(f"the validation score is {valid_score:3f}")
    print(f'the mse is {mse:3f}')
    print(f"the r2 score is: {r2:3f}")
    return y_pred

def train_models(x, y, random_state=0):
    x_train, x_valid, y_train, y_valid = train_test_split(x, y, 
                                                        random_state=random_state, 
                                                        test_size=0.1)    
    y_pred = do_rf(x_train, y_train, x_valid, y_valid)
    _ = do_xgb(x_train, y_train, x_valid, y_valid)
    return y_pred

### Lets start with our Buffer circuit

_ = train_models(X_buffer, y_buffer/y_nor.min())

### Inverter

_ = train_models(X_inverter, y_inverter/y_inverter.min(), random_state=0)

### Nand

_ = train_models(X_nand, y_nand/y_nand.min(), random_state=1)

### Nor

_ = train_models(X_nor, y_nor/y_nor.min())
