==> 05_cross_gate_init.ipynb <==
### Initial work exploring whether we can train generalizable models --> i.e. given training data on `and` gates and `nor` gates, can we make accurate predictions on `Nand` gates?

#### will stick with the simple models for now while fleshing it out. Can introduce other types of model/algo later on

from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import Ridge, RidgeCV
from sklearn.ensemble import RandomForestRegressor 
from sklearn.metrics import mean_squared_error, r2_score
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import PolynomialFeatures
from sklearn.linear_model import ElasticNet, LassoCV, MultiTaskLassoCV, MultiTaskElasticNetCV
from sklearn import linear_model
import matplotlib.pyplot as plt
import pandas as pd
import numpy as np
import xgboost as xgb

First, we'll extract the data for a single run. This should feature 10 different gates

df = pd.read_csv('data/ten_gate.csv')
df.head(2)

df.voltage.unique()

df.thickness.unique()

df.clock_cycle.unique()

unique_gates = list(df.gate_type.unique())
np.array(unique_gates)

We'll need to drop some failed values in our nand4 gate in order to proceed, you can see the heatmap at the bottom of notebook #3 in order to confirm the need for this

#these are the bad t_fall values in nand4 that need to be dropped
nand4_drop_idxs = ((df['gate_type'] == 'nand4') & (df['t_fall'] == ' FAILED')).values
nand4_drop_idxs.sum()

df = df.loc[~nand4_drop_idxs].reset_index(drop=True)
df.shape

Lets flesh out a few functions that will make our lives easier in order to index into our data appropriately and drop whatever items we might need to be removed

def get_gate(gate_type='buffer'): return df[df.gate_type == gate_type]

def get_gate_xy(gate_type='buffer'):
    #y_drop = 't_fall' if gate_type == 'buffer' else 't_rise'
    df = get_gate(gate_type)
    x_cols = ['clock_cycle', 'thickness', 'voltage']
    y_cols = ['t_delay', 't_rise', 't_fall']
    #y_cols.remove(y_drop)
    x = np.vstack(df[x_cols].values).astype('float')
    y = np.vstack(np.vstack(df[y_cols].values)).astype('float')
    return x, y

x1, y1 = get_gate_xy('nor3')
x2, y2 = get_gate_xy('nand2')
x3, y3 = get_gate_xy('buffer')
x1.shape, y1.shape, x2.shape, y2.shape, x3.shape, x3.shape

gate_dict = dict.fromkeys(unique_gates)
gate_dict

for gate in unique_gates: 
    gate_dict[gate] = get_gate_xy(gate_type=gate)

x, y = gate_dict['buffer']
x_tr, x_val, y_tr, y_val = train_test_split(x,y)
x_tr.shape, x_val.shape, y_tr.shape, y_val.shape

def get_mult_gate_df(gate_names=['buffer']):
    gates = [gate_dict[gate] for gate in gate_names]
    xs = np.vstack([gate[0] for gate in gates])
    ys = np.vstack([gate[1] for gate in gates])
    #xs = np.append(*xs, axis=0)
    #ys = np.append(*ys, axis=0)
    return xs, ys

xs_tr, ys_tr = get_mult_gate_df(['nor2', 'nand2'])
x_val, y_val = get_mult_gate_df(['buffer'])
xs_tr.shape, ys_tr.shape, x_val.shape, y_val.shape

#lets introduce partial functions to clean up our code a little
from functools import partial

def normalize(x): return x / x.min() #used to normalize our Y values

def poly_x(x, num_poly=2): return PolynomialFeatures(num_poly).fit_transform(x)

def do_simple_model(x_tr, y_tr, x_val, y_val, model_name='ridge', num_poly=2, alpha=1.0, norm_y=True, print_vals=True,
                   get_model=False):
    #x_tr, x_val, y_tr, y_val = train_test_split(x,y)
    
    #lets convert our independent variable into a different polynomial combination
    if num_poly: x_tr, x_val = poly_x(x_tr, num_poly), poly_x(x_val, num_poly)
    
    alps = np.logspace(-5,-1)
    
    if model_name   == 'ridge': model = Ridge(alpha=alpha, random_state=0)
    elif model_name == 'ridge-cv': model = RidgeCV()#alphas=alps)
    #elif model_name == 'lasso': model = linear_model.Lasso(alpha=alpha,random_state=0)
    #elif model_name == 'elastic': model = ElasticNet(random_state=0)
    elif model_name == 'lasso': model = MultiTaskLassoCV(random_state=0, alphas=alps, max_iter=10000)
    elif model_name == 'elastic': model = MultiTaskElasticNetCV(random_state=0, alphas=alps, max_iter=10000,
                                                                  l1_ratio=np.arange(0.1, 1, 0.1))
    elif model_name == 'random_forest': model = RandomForestRegressor(max_depth=100, random_state=0)
    elif model_name == 'xgboost': model = xgb.XGBRegressor(eval_metric='rmse', random_state=0)
    
    #normally want to normalize y values for regression problem
    if norm_y: y_tr, y_val = normalize(y_tr), normalize(y_val)
    
    #now lets fit our model to our training data
    model.fit(x_tr, y_tr)
    
    #we'll predict our R2 score on only the validation data
    r2score = model.score(x_val, y_val)
    if print_vals: print(f'for a {model_name} model, the validation R2 score is: {r2score:2f}')
    if get_model: return model
    return r2score

get_mult_gate_df([gates[-1]])

def run_models(gate_types=['buffer','nand2', 'nor3', 'nor4', 'nor2'], random_state=0, print_vals=True):
    print(f'-----Training with {gate_types[:-1]} gate(s)-----')
    print(f'-----Validating with {gate_types[-1]} gate(s)-----')
    x_tr, y_tr = get_mult_gate_df(gate_types[:-1])
    x_val, y_val = get_mult_gate_df([gate_types[-1]])
    scores = []
    do_model = partial(do_simple_model, x_tr, y_tr,x_val, y_val, print_vals=print_vals)
    #do_model(model_name='lasso')
    #do_model(model_name='elastic')
    scores.append(do_model(model_name='ridge', norm_y=False))
    #scores.append(do_model(model_name='ridge-cv', norm_y=True))
    scores.append(do_model(model_name='lasso'))
    scores.append(do_model(model_name='elastic'))
    scores.append(do_model(model_name='random_forest', num_poly=False))
    scores.append(do_model(model_name='xgboost', num_poly=False))
    return scores
    print('\n')

#test run on just the buffer gate
_ = run_models(print_vals=True)

_ = run_models(print_vals=True, gate_types=['nor2', 'nor3', 'nor4'])

_ = run_models(print_vals=True, gate_types=['nor2', 'nor3', 'buffer', 'nor4'])

_ = run_models(print_vals=True, gate_types=['nor2', 'nor3', 'nor4', 'buffer'])

_ = run_models(print_vals=True, gate_types=['nand2', 'nand3', 'nand4'])

#list(gate_dict.keys())

_ = run_models(print_vals=True, gate_types=list(gate_dict.keys()))
